Added supervisor and changed evaluator to supervisor

+++ app/Config/Routes.php
@@ -11,25 +11,28 @@ $routes->post('/test/importWordTable', 'Test::importWordTable');
 
 $routes->group('', ['filter' => 'auth'], function($routes) {
     //Ratings
-    $routes->get('ratings', 'Rating', ['filter' => 'admin']);
+    $routes->get('ratings', 'Rating', ['filter' => 'role:admin,supervisor']);
+    $routes->get('rating/departments', 'Rating::departments', ['filter' => 'role:admin']);
+    $routes->get('rating/show', 'Rating::show', ['filter' => 'role:admin,supervisor']);
+    $routes->post('rating/save', 'Rating::save', ['filter' => 'role:admin,supervisor']);
 
-    $routes->get('rating/departments', 'Rating::departments', ['filter' => 'admin']);
-    $routes->get('rating/show', 'Rating::show', ['filter' => 'admin']);
-    $routes->delete('rating', 'Rating::destroy', ['filter' => 'admin']);
-    $routes->post('rating/save', 'Rating::save', ['filter' => 'admin']);
+    $routes->delete('rating', 'Rating::destroy', ['filter' => 'role:admin']);
+    $routes->post('rating/save', 'Rating::save', ['filter' => 'role:admin']);
 
     $routes->get('profile', 'Profile');
 
     // Document
-    $routes->get('documents', 'Document');
+    $routes->get('documents', 'Folder');
+    $routes->get('folder', 'Folder::show');
+    $routes->post('folder', 'Folder::store', ['filter' => 'role:admin']);
+    $routes->post('folder/send', 'Folder::send', ['filter' => 'role:admin']);
+    $routes->delete('folder', 'Folder::destroy', ['filter' => 'role:admin']);
 
     $routes->get('document', 'Document::show');
     $routes->patch('document', 'Document::update');
     $routes->post('document', 'Document::store');
     $routes->delete('document', 'Document::destroy');
     $routes->post('document/share', 'Document::share');
-    $routes->post('document/send', 'Document::send', ['filter' => 'admin']);
-    $routes->post('document/createFolder', 'Document::createFolder', ['filter' => 'admin']);
     $routes->get('document/count', 'Document::count');
 
     // Submission
